pyproject.toml
[tool.poetry]
name = "curb_backend"
version = "0.1.0"
description = ""
authors = ["KimDuHong <art970@naver.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
django-environ = "^0.10.0"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.14.0"
gunicorn = "^20.1.0"
sentry-sdk = "^1.19.1"
psycopg2-binary = "^2.9.6"
Faker = "^18.4.0"
PyKakao = "^0.0.4"
django-debug-toolbar = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
